+++ Tests/SpecFlow.VisualStudio.Specs/StepDefinitions/ProjectSystemSteps.cs
@@ -692,9 +692,8 @@ namespace SpecFlow.VisualStudio.Specs.StepDefinitions
                 });
             WhenIInvokeTheCommand(_commandToInvokeDeferred);
 
-            var cts = new CancellationTokenSource(TimeSpan.FromSeconds(1));
             await _projectScope.StubIdeScope.AnalyticsTransmitter
-                .WaitForEventAsync("Rename step command executed", cts.Token);
+                .WaitForEventAsync("Rename step command executed");
         }
 
         [Then("invoking the first item from the jump list renames the {string} {string} step definition")]
@@ -710,9 +709,8 @@ namespace SpecFlow.VisualStudio.Specs.StepDefinitions
 
             InvokeFirstContextMenuItem();
 
-            var cts = new CancellationTokenSource(TimeSpan.FromSeconds(1));
             await _projectScope.StubIdeScope.AnalyticsTransmitter
-                .WaitForEventAsync("Rename step command executed", cts.Token);
+                .WaitForEventAsync("Rename step command executed");
 
             string fileContent = _wpfTextView.TextSnapshot.GetText();
             var parsedSnippets = ParseSnippetsFromFile(fileContent);