+++ Tests/SpecFlow.VisualStudio.VsxStubs/ProjectSystem/StubAnalyticsTransmitter.cs
@@ -1,6 +1,8 @@
-﻿using System;
+﻿#nullable enable
+using System;
 using System.Collections.Concurrent;
 using System.Collections.Generic;
+using System.Diagnostics;
 using System.Linq;
 using System.Threading;
 using System.Threading.Tasks;
@@ -11,18 +13,27 @@ namespace SpecFlow.VisualStudio.VsxStubs.ProjectSystem
     public class StubAnalyticsTransmitter : IAnalyticsTransmitter
     {
         private ConcurrentBag<IAnalyticsEvent> Events { get; } =new ();
-        private readonly TaskCompletionSource<IAnalyticsEvent> _taskCompletionSource = new();
+        private TaskCompletionSource<IAnalyticsEvent> _transmitSignal = new();
 
         public void TransmitEvent(IAnalyticsEvent runtimeEvent)
         {
             Events.Add(runtimeEvent);
-            _taskCompletionSource.TrySetResult(runtimeEvent);
+            var signal = Interlocked.Exchange(ref _transmitSignal, new TaskCompletionSource<IAnalyticsEvent>());
+            signal.SetResult(runtimeEvent);
         }
 
         public void TransmitExceptionEvent(Exception exception, Dictionary<string, object> additionalProps = null, bool? isFatal = null,
             bool anonymize = true)
         {
            //nop
+        }
+
+        public Task<IAnalyticsEvent> WaitForEventAsync(string eventName)
+        {
+            CancellationTokenSource cts = Debugger.IsAttached
+                ? new CancellationTokenSource(TimeSpan.FromMinutes(1))
+                : new CancellationTokenSource(TimeSpan.FromSeconds(2));
+            return WaitForEventAsync(eventName, cts.Token);
         }
 
         public async Task<IAnalyticsEvent> WaitForEventAsync(string eventName, CancellationToken cancellationToken)
@@ -32,7 +43,7 @@ namespace SpecFlow.VisualStudio.VsxStubs.ProjectSystem
                 var analyticsEvent = Events.FirstOrDefault(ev => ev.EventName == eventName);
                 if (analyticsEvent != null) return analyticsEvent;
 
-                await Task.WhenAny(_taskCompletionSource.Task, Task.Delay(-1, cancellationToken));
+                await Task.WhenAny(_transmitSignal.Task, Task.Delay(-1, cancellationToken));
             }
 
             throw new TaskCanceledException();